Move forge/OF installer stuff to a different package and fix deep nesting of if statements in the window handler
