+++ app_pojavlauncher/src/main/assets/components/forge_installer/version
@@ -1,1 +1,1 @@
-1688118307460
+1688133008591
