+++ app_pojavlauncher/src/main/jni/stdio_is.c
@@ -85,7 +85,7 @@ Java_net_kdt_pojavlaunch_Logger_begin(JNIEnv *env, __attribute((unused)) jclass 
 
     /* open latestlog.txt for writing */
     const char* logFilePath = (*env)->GetStringUTFChars(env, logPath, NULL);
-    latestlog_fd = open(logFilePath, O_WRONLY | O_TRUNC | O_CREAT, 644);
+    latestlog_fd = open(logFilePath, O_WRONLY | O_TRUNC);
     if(latestlog_fd == -1) {
         latestlog_fd = 0;
         (*env)->ThrowNew(env, ioeClass, strerror(errno));
