+++ forge_installer/src/main/java/git/artdeell/installer_agent/MainWindowFilter.java
@@ -1,4 +1,4 @@
-package git.artdeell.forgeinstaller;
+package git.artdeell.installer_agent;
 
 import javax.swing.*;
 import java.awt.*;
